satdata_client/satdata_client/requirements.txt
requests
numpy
Pillow
sentinelhub
